bcftbx.utils: relocate 'split_into_lines' into new 'utils.text' submodule.
